refactor(#2400): address PR review comments (jsonb atomic append, validateBody deduplication)
