+++ server/src/module/skill-test/skill-test.routes.ts
@@ -3,7 +3,8 @@ import { SkillTestService } from "./skill-test.service.js";
 import { SkillTestController } from "./skill-test.controller.js";
 import { authMiddleware } from "../../middleware/auth.middleware.js";
 import { requireRole } from "../../middleware/role.middleware.js";
-import { validateBody, proctorLogBatchSchema } from "./skill-test.validation.js";
+import { validateBody } from "../../middleware/validation.middleware.js";
+import { proctorLogBatchSchema } from "./skill-test.validation.js";
 
 const service = new SkillTestService();
 const controller = new SkillTestController(service);
